backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.9
google-genai>=1.0.0
python-jobspy
pandas>=2.0.0
reportlab==4.2.0
pydantic==2.7.1
httpx==0.27.0
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
apscheduler==3.10.4
email-validator>=2.0.0
pdfminer.six>=20221105
pypdf>=4.0.0
slowapi>=0.1.9
bcrypt>=4.0.0
